✨ <feat> add files for feedback form.

+++ src/main/java/projetb2/overlook_hotel/service/BookingService.java
@@ -35,7 +35,7 @@ public class BookingService {
      * If the booking is successfully accepted, it returns true.
      */
     @Transactional
-    public boolean acceptBooking(Long bookingId) {
+    public boolean acceptBooking(Integer bookingId) {
         return bookingRepo.findById(bookingId).map(booking -> {
             if (booking.getBookingStatus() != BookingStatus.PENDING)
                 return false;
@@ -50,7 +50,7 @@ public class BookingService {
      * If the booking is successfully declined, it returns true.
      */
     @Transactional
-    public boolean declineBooking(Long bookingId) {
+    public boolean declineBooking(Integer bookingId) {
         return bookingRepo.findById(bookingId).map(booking -> {
             if (booking.getBookingStatus() != BookingStatus.PENDING)
                 return false;
@@ -65,7 +65,12 @@ public class BookingService {
      * If the booking is successfully updated, it returns true.
      */
     @Transactional
-    public boolean updateBooking(Long bookingId, Long roomId, LocalDate arrival, LocalDate departure, BookingStatus status) {
+    public boolean updateBooking(
+        Integer bookingId,
+        Integer roomId,
+        LocalDate arrival,
+        LocalDate departure,
+        BookingStatus status) {
         return bookingRepo.findById(bookingId).map(booking -> {
             Room room = roomService.getRoomById(roomId);
             if (room == null)
